# dockstore-webservice/src/main/resources/openapi3/openapi.yaml
components:
  schemas:
    Alias:
      properties:
        content:
          type: string
      type: object
    Aliasable:
      properties:
        aliases:
          additionalProperties:
            $ref: '#/components/schemas/Alias'
          type: object
      type: object
...
    CollectionOrganization:
      properties:
        collectionDisplayName:
          type: string
        collectionId:
          format: int64
          type: integer
        collectionName:
          type: string
        organizationDisplayName:
          type: string
        organizationId:
          format: int64
          type: integer
        organizationName:
          type: string
      type: object
...
    Entry:
      properties:
        aliases:
          additionalProperties:
            $ref: '#/components/schemas/Alias'
          type: object
        author:
          type: string
        checker_id:
          format: int64
          type: integer
        dbCreateDate:
          format: date-time
          type: string
        dbUpdateDate:
          format: date-time
          type: string
        defaultVersion:
          type: string
        description:
          type: string
        email:
          type: string
        gitUrl:
          type: string
        has_checker:
          type: boolean
        id:
          format: int64
          type: integer
        input_file_formats:
          items:
            $ref: '#/components/schemas/FileFormat'
          type: array
          uniqueItems: true
        is_published:
          type: boolean
        labels:
          items:
            $ref: '#/components/schemas/Label'
          type: array
          uniqueItems: true
        lastUpdated:
          format: date-time
          type: string
        last_modified:
          format: int32
          type: integer
        last_modified_date:
          format: date-time
          type: string
        output_file_formats:
          items:
            $ref: '#/components/schemas/FileFormat'
          type: array
          uniqueItems: true
        starredUsers:
          items:
            $ref: '#/components/schemas/User'
          type: array
          uniqueItems: true
        topicId:
          format: int64
          type: integer
        users:
          items:
            $ref: '#/components/schemas/User'
          type: array
          uniqueItems: true
        workflowVersions:
          items:
            $ref: '#/components/schemas/Version'
          type: array
          uniqueItems: true
      type: object
...
    FileFormat:
      properties:
        value:
          type: string
      type: object
    Label:
      properties:
        id:
          format: int64
          type: integer
        value:
          type: string
      type: object
...
    Organization:
      properties:
        aliases:
          additionalProperties:
            $ref: '#/components/schemas/Alias'
          type: object
        avatarUrl:
          pattern: ([^\s]+)(?i)(\.jpg|\.jpeg|\.png|\.gif)
          type: string
        dbCreateDate:
          format: date-time
          type: string
        dbUpdateDate:
          format: date-time
          type: string
        description:
          type: string
        displayName:
          maxLength: 50
          minLength: 3
          pattern: '[\w ,_\-&()'']*'
          type: string
        email:
          type: string
        id:
          format: int64
          type: integer
        link:
          type: string
        location:
          type: string
        name:
          maxLength: 39
          minLength: 3
          pattern: '[a-zA-Z][a-zA-Z\d]*'
          type: string
        starredUsers:
          items:
            $ref: '#/components/schemas/User'
          type: array
          uniqueItems: true
        status:
          enum:
          - PENDING
          - REJECTED
          - APPROVED
          type: string
        topic:
          type: string
        users:
          items:
            $ref: '#/components/schemas/OrganizationUser'
          type: array
          uniqueItems: true
      type: object
    OrganizationUser:
      properties:
        accepted:
          type: boolean
        dbCreateDate:
          format: date-time
          type: string
        dbUpdateDate:
          format: date-time
          type: string
        id:
          $ref: '#/components/schemas/OrganizationUserId'
        organization:
          $ref: '#/components/schemas/Organization'
        role:
          enum:
          - MAINTAINER
          - MEMBER
          type: string
        user:
          $ref: '#/components/schemas/User'
      type: object
    OrganizationUserId:
      properties:
        organizationId:
          format: int64
          type: integer
        userId:
          format: int64
          type: integer
      type: object
...
    Profile:
      properties:
        avatarURL:
          type: string
        bio:
          type: string
        company:
          type: string
        email:
          type: string
        location:
          type: string
        name:
          type: string
        username:
          type: string
      type: object
...
    SourceFile:
      properties:
        absolutePath:
          type: string
        content:
          type: string
        frozen:
          type: boolean
        id:
          format: int64
          type: integer
        path:
          type: string
        type:
          enum:
          - DOCKSTORE_CWL
          - DOCKSTORE_WDL
          - DOCKERFILE
          - CWL_TEST_JSON
          - WDL_TEST_JSON
          - NEXTFLOW
          - NEXTFLOW_CONFIG
          - NEXTFLOW_TEST_PARAMS
          - DOCKSTORE_YML
          - DOCKSTORE_SERVICE_YML
          - DOCKSTORE_SERVICE_TEST_JSON
          - DOCKSTORE_SERVICE_OTHER
          type: string
        verifiedBySource:
          additionalProperties:
            $ref: '#/components/schemas/VerificationInformation'
          type: object
      type: object
...
    User:
      properties:
        avatarUrl:
          type: string
        curator:
          type: boolean
        id:
          format: int64
          type: integer
        isAdmin:
          type: boolean
        name:
          type: string
        privacyPolicyVersion:
          enum:
          - NONE
          - PRIVACY_POLICY_VERSION_2_5
          type: string
        privacyPolicyVersionAcceptanceDate:
          format: date-time
          type: string
        setupComplete:
          type: boolean
        tosacceptanceDate:
          format: date-time
          type: string
        tosversion:
          enum:
          - NONE
          - TOS_VERSION_1
          type: string
        tosversionAcceptanceDate:
          format: date-time
          type: string
          writeOnly: true
        userProfiles:
          additionalProperties:
            $ref: '#/components/schemas/Profile'
          type: object
        username:
          type: string
      type: object
    Validation:
      properties:
        id:
          format: int64
          type: integer
        message:
          type: string
        type:
          enum:
          - DOCKSTORE_CWL
          - DOCKSTORE_WDL
          - DOCKERFILE
          - CWL_TEST_JSON
          - WDL_TEST_JSON
          - NEXTFLOW
          - NEXTFLOW_CONFIG
          - NEXTFLOW_TEST_PARAMS
          - DOCKSTORE_YML
          - DOCKSTORE_SERVICE_YML
          - DOCKSTORE_SERVICE_TEST_JSON
          - DOCKSTORE_SERVICE_OTHER
          type: string
        valid:
          type: boolean
      type: object
    VerificationInformation:
      properties:
        metadata:
          type: string
        platformVersion:
          type: string
        verified:
          type: boolean
      type: object
    Version:
      properties:
        commitID:
          type: string
        dbUpdateDate:
          format: date-time
          type: string
        dirtyBit:
          type: boolean
        doiStatus:
          enum:
          - NOT_REQUESTED
          - REQUESTED
          - CREATED
          type: string
        doiURL:
          type: string
        frozen:
          type: boolean
        hidden:
          type: boolean
        id:
          format: int64
          type: integer
        input_file_formats:
          items:
            $ref: '#/components/schemas/FileFormat'
          type: array
          uniqueItems: true
        name:
          type: string
        output_file_formats:
          items:
            $ref: '#/components/schemas/FileFormat'
          type: array
          uniqueItems: true
        reference:
          type: string
        referenceType:
          enum:
          - COMMIT
          - TAG
          - BRANCH
          - NOT_APPLICABLE
          - UNSET
          type: string
        sourceFiles:
          items:
            $ref: '#/components/schemas/SourceFile'
          type: array
          uniqueItems: true
        valid:
          type: boolean
        validations:
          items:
            $ref: '#/components/schemas/Validation'
          type: array
          uniqueItems: true
        verified:
          type: boolean
        verifiedSource:
          type: string
        verifiedSources:
          items:
            type: string
          type: array
        versionEditor:
          $ref: '#/components/schemas/User'
        workingDirectory:
          type: string
      type: object
    Workflow:
      discriminator:
        propertyName: type
      properties:
        aliases:
          additionalProperties:
            $ref: '#/components/schemas/Alias'
          type: object
        author:
          type: string
        checker_id:
          format: int64
          type: integer
        dbCreateDate:
          format: date-time
          type: string
        dbUpdateDate:
          format: date-time
          type: string
        defaultTestParameterFilePath:
          type: string
        defaultVersion:
          type: string
        description:
          type: string
        descriptorType:
          enum:
          - CWL
          - WDL
          - NFL
          - service
          - cwl
          - wdl
          type: string
        email:
          type: string
        full_workflow_path:
          type: string
        gitUrl:
          type: string
        has_checker:
          type: boolean
        id:
          format: int64
          type: integer
        input_file_formats:
          items:
            $ref: '#/components/schemas/FileFormat'
          type: array
          uniqueItems: true
        isChecker:
          type: boolean
        is_published:
          type: boolean
        labels:
          items:
            $ref: '#/components/schemas/Label'
          type: array
          uniqueItems: true
        lastUpdated:
          format: date-time
          type: string
        last_modified:
          format: int32
          type: integer
        last_modified_date:
          format: date-time
          type: string
        mode:
          enum:
          - FULL
          - STUB
          - HOSTED
          - SERVICE
          type: string
        organization:
          type: string
        output_file_formats:
          items:
            $ref: '#/components/schemas/FileFormat'
          type: array
          uniqueItems: true
        parentEntry:
          $ref: '#/components/schemas/Entry'
        path:
          type: string
        repository:
          type: string
        sourceControl:
          enum:
          - dockstore.org
          - github.com
          - bitbucket.org
          - gitlab.com
          type: string
        source_control_provider:
          type: string
        starredUsers:
          items:
            $ref: '#/components/schemas/User'
          type: array
          uniqueItems: true
        topicId:
          format: int64
          type: integer
        users:
          items:
            $ref: '#/components/schemas/User'
          type: array
          uniqueItems: true
        workflowName:
          type: string
        workflowVersions:
          items:
            $ref: '#/components/schemas/WorkflowVersion'
          type: array
          uniqueItems: true
        workflow_path:
          type: string
      type: object
    WorkflowVersion:
      properties:
        commitID:
          type: string
        dbUpdateDate:
          format: date-time
          type: string
        dirtyBit:
          type: boolean
        doiStatus:
          enum:
          - NOT_REQUESTED
          - REQUESTED
          - CREATED
          type: string
        doiURL:
          type: string
        frozen:
          type: boolean
        hidden:
          type: boolean
        id:
          format: int64
          type: integer
        input_file_formats:
          items:
            $ref: '#/components/schemas/FileFormat'
          type: array
          uniqueItems: true
        last_modified:
          format: date-time
          type: string
        name:
          type: string
        output_file_formats:
          items:
            $ref: '#/components/schemas/FileFormat'
          type: array
          uniqueItems: true
        reference:
          type: string
        referenceType:
          enum:
          - COMMIT
          - TAG
          - BRANCH
          - NOT_APPLICABLE
          - UNSET
          type: string
        sourceFiles:
          items:
            $ref: '#/components/schemas/SourceFile'
          type: array
          uniqueItems: true
        subClass:
          enum:
          - DOCKER_COMPOSE
          - SWARM
          - KUBERNETES
          - HELM
          type: string
        valid:
          type: boolean
        validations:
          items:
            $ref: '#/components/schemas/Validation'
          type: array
          uniqueItems: true
        verified:
          type: boolean
        verifiedSource:
          type: string
        verifiedSources:
          items:
            type: string
          type: array
        versionEditor:
          $ref: '#/components/schemas/User'
        workflow_path:
          type: string
        workingDirectory:
          type: string
      type: object
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  contact:
    email: theglobalalliance@genomicsandhealth.org
    name: Dockstore@ga4gh
    url: https://discuss.dockstore.org/t/opening-helpdesk-tickets/1506
  description: The Dockstore API, includes proprietary and GA4GH V1+V2 endpoints
  license:
    name: Apache 2.0
    url: https://github.com/dockstore/dockstore/blob/develop/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Dockstore
  version: 1.6.0
openapi: 3.0.1
paths:
...
  /entries/{id}/aliases:
    put:
      operationId: updateAliases_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
      - in: query
        name: aliases
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
          description: default response
  /entries/{id}/collections:
    get:
      operationId: entryCollections
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
      responses:
        default:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CollectionOrganization'
                type: array
          description: default response
  /entries/{id}/topic:
    post:
      description: Create a discourse topic for an entry.
      operationId: setDiscourseTopic
      parameters:
      - description: The id of the entry to add a topic to.
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
          description: default response
      security:
      - bearer: []
...
  /organizations/collections/{collectionId}/aliases:
    put:
      operationId: updateAliases_1
      parameters:
      - in: path
        name: collectionId
        required: true
        schema:
          format: int64
          type: integer
      - in: query
        name: aliases
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aliasable'
          description: default response
...
  /organizations/{organizationId}/aliases:
    put:
      operationId: updateAliases_3
      parameters:
      - in: path
        name: organizationId
        required: true
        schema:
          format: int64
          type: integer
      - in: query
        name: aliases
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: default response
...
  /workflows/{alias}/aliases:
    get:
      operationId: getWorkflowByAlias
      parameters:
      - in: path
        name: alias
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
          description: default response
...